Fix some functions.
In `browse-at-remote/origin`, `memq` was used to test the remote list
for an `"origin"` entry. This doesn’t work because it compares object
identity instead of value, and strings aren’t interned. Change to
`member`, simplify logic, and add an error if there are no remotes.

Also, `browse-at-remote/get-remotes` would return a list containing an
empty string for repos with no remotes. This has been fixed, it now
returns nil for these repos.